wip: move route options to declaration
