+++ src/routing/declaration/route-options.type.ts
