+++ src/routing/declaration/route-props.type.ts
@@ -1,6 +1,6 @@
 import type { HttpMiddleware, HttpRequest } from '@/Http';
 import type { HttpMethod } from '@/routing/declaration/http-method.type';
-import type { RouteOptions } from '@/routing/route/route-options.type';
+import type { RouteOptions } from '@/routing/declaration/route-options.type';
 import type { Request } from 'koa';
 
 export interface RouteProps<TRequest extends Request = HttpRequest> {
