+++ src/routing/deprecated-decorator/route.ts
@@ -1,6 +1,6 @@
 import type { HttpMiddleware } from '@/Http';
 import type { HttpMethod } from '@/routing/declaration/http-method.type';
-import type { RouteOptions } from '@/routing/route/route-options.type';
+import type { RouteOptions } from '@/routing/declaration/route-options.type';
 import { createLegacyRouteDecorator } from './legacy-router';
 
 /**
