+++ src/routing/normalization/normalize-route-props.ts
@@ -1,8 +1,8 @@
 import type { HttpRequest } from '@/Http';
 import type { HttpMethod } from '@/routing/declaration/http-method.type';
 import type { NormalizedRouteProps } from '@/routing/declaration/normalized-route-props.type';
+import type { RouteOptions } from '@/routing/declaration/route-options.type';
 import { RouteProps } from '@/routing/declaration/route-props.type';
-import type { RouteOptions } from '@/routing/route/route-options.type';
 import type { RouterMethod } from '@/routing/route/router-method.type';
 import type { Request } from 'koa';
